embedder/requirements.txt
sentence-transformers
psycopg2-binary
numpy
pandas
sqlalchemy